requirements.txt
streamlit
python-dotenv
pypdf
langchain-text-splitters
sentence-transformers
pinecone
neo4j
